changed gain parameter to double type in setGain(idx, gain)
git-svn-id: https://svn.artisynth.org/svn/artisynth_core/trunk@309 7661e47b-f193-4b25-995f-cdbae50cbfef

+++ src/artisynth/core/mechmodels/MuscleExciter.java
@@ -261,7 +261,7 @@ ExcitationComponent {
       return myTargets.get(idx).getExcitationGain (this);
    }
 
-   public void setGain (int idx, int gain) {
+   public void setGain (int idx, double gain) {
       if (idx >= myTargets.size()) {
          throw new ArrayIndexOutOfBoundsException ("index " + idx
          + " is out of bounds");
